dom/bindings/test/test_proxy_expandos.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=965992
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 965992</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=965992">Mozilla Bug 965992</a>
<p id="display"></p>
<form id="theform"></form>
<pre id="test">
<script type="application/javascript">

// Ensure we are in JIT code and attach IC stubs.
const iterations = 50;

function testFoo(obj, kind, expected) {
  for (var i = 0; i < iterations; i++) {
    obj.foo = i;
    is(obj.foo, (expected === undefined) ? i : expected,
       "Looking up an expando should work - " + kind);
  }
}

function getPropTests(obj) {
  // Start with a plain data property.
  obj.foo = "bar";
  testFoo(obj, "plain");

  // Now change it to a scripted getter/setter.
  var count = 0;
  var getterSetterVal = 0;
  Object.defineProperty(obj, "foo", {configurable: true, get() {
    is(this, obj, "Getter should have the proxy as |this|");
    is(arguments.length, 0, "Shouldn't pass arguments to getters");
    count++;
    return getterSetterVal;
  }, set(v) {
    is(this, obj, "Setter should have the proxy as |this|");
    is(arguments.length, 1, "Should pass 1 argument to setters");
    getterSetterVal = v;
    count++;
  }});
  testFoo(obj, "scripted getter/setter");
  is(count, iterations * 2, "Should have called the getter/setter enough times");

  // Now try a native getter/setter.
  Object.defineProperty(obj, "foo", {get: Math.abs, set: Math.abs, configurable: true});
  testFoo(obj, "native getter/setter", NaN);
}

function getElemTests(obj) {
  // Define two expando properties, then test inline caches for obj[prop]
  // correctly guard on prop being the same.
  var count = 0, getterSetterVal = 0;
  obj.elem1 = 1;
  Object.defineProperty(obj, "elem2", {
                        get() { count++; return getterSetterVal; },
                        set(v) { getterSetterVal = v; count++; },
  });
  for (var i = 0; i < iterations; i++) {
    var prop = ((i & 1) == 0) ? "elem1" : "elem2";
    obj[prop] = i;
    is(obj[prop], i, "Should return correct property value");
  }
  is(count, iterations, "Should have called the getter/setter enough times");
}

var directExpando = document.getElementsByTagName("*");
var indirectExpando = document.getElementById("theform");

getPropTests(directExpando);
getPropTests(indirectExpando);

getElemTests(indirectExpando);
getElemTests(directExpando);

</script>
</pre>
</body>
</html>
